chore: move all dsn logic to internal/dsn
