+++ backend/controller/sql/database_integration_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/alecthomas/assert/v2"
 
-	"github.com/TBD54566975/ftl/backend/controller/dsn"
+	"github.com/TBD54566975/ftl/internal/dsn"
 	in "github.com/TBD54566975/ftl/internal/integration"
 )
