+++ internal/dsn/dsn.go
@@ -1,6 +1,11 @@
 package dsn
 
-import "fmt"
+import (
+	"context"
+	"fmt"
+
+	"github.com/TBD54566975/ftl/internal/schema"
+)
 
 type dsnOptions struct {
 	host string
@@ -37,4 +42,20 @@ func MySQLDSN(dbName string, options ...Option) string {
 		opt(opts)
 	}
 	return fmt.Sprintf("root:secret@tcp(%s:%d)/%s?allowNativePasswords=True", opts.host, opts.port, dbName)
+}
+
+func ResolvePostgresDSN(ctx context.Context, connector schema.DatabaseConnector) (string, error) {
+	dsnRuntime, ok := connector.(*schema.DSNDatabaseConnector)
+	if !ok {
+		return "", fmt.Errorf("unexpected database connector type: %T", connector)
+	}
+	return dsnRuntime.DSN, nil
+}
+
+func ResolveMySQLDSN(ctx context.Context, connector schema.DatabaseConnector) (string, error) {
+	dsnRuntime, ok := connector.(*schema.DSNDatabaseConnector)
+	if !ok {
+		return "", fmt.Errorf("unexpected database connector type: %T", connector)
+	}
+	return dsnRuntime.DSN, nil
 }
