+++ internal/integration/actions.go
@@ -28,9 +28,9 @@ import (
 
 	"github.com/block/scaffolder"
 
-	"github.com/TBD54566975/ftl/backend/controller/dsn"
 	ftlv1 "github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1"
 	schemapb "github.com/TBD54566975/ftl/backend/protos/xyz/block/ftl/v1/schema"
+	"github.com/TBD54566975/ftl/internal/dsn"
 	ftlexec "github.com/TBD54566975/ftl/internal/exec"
 	"github.com/TBD54566975/ftl/internal/log"
 )
